pyproject.toml
[project]
name = "jp-to-cn-subtitle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "git-filter-repo>=2.47.0",
    "gradio>=5.49.1",
    "openai>=2.3.0",
    "openai-whisper>=20250625",
    "python-dotenv>=1.1.1",
    "srt>=3.5.3",
    "tqdm>=4.67.1",
]
